Clean up

+++ space-age/src/lib.rs
@@ -11,13 +11,13 @@ macro_rules! planet {
 }
 
 const EARTH     : f64 = 31557600.0;
-const MERCURY   : f64 = 0.2408467  * EARTH as f64;
-const VENUS     : f64 = 0.61519726 * EARTH as f64;
-const MARS      : f64 = 1.8808158  * EARTH as f64;
-const JUPITER   : f64 = 11.862615  * EARTH as f64;
-const SATURN    : f64 = 29.447498  * EARTH as f64;
-const URANUS    : f64 = 84.016846  * EARTH as f64;
-const NEPTUNE   : f64 = 164.79132  * EARTH as f64;
+const MERCURY   : f64 = 0.2408467  * EARTH;
+const VENUS     : f64 = 0.61519726 * EARTH;
+const MARS      : f64 = 1.8808158  * EARTH;
+const JUPITER   : f64 = 11.862615  * EARTH;
+const SATURN    : f64 = 29.447498  * EARTH;
+const URANUS    : f64 = 84.016846  * EARTH;
+const NEPTUNE   : f64 = 164.79132  * EARTH;
 
 pub type Duration = f64;
